requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu